+++ ContentPatcher/Framework/Migrations/Migration_2_0.ForNpcDispositions.cs
@@ -27,6 +27,9 @@ namespace ContentPatcher.Framework.Migrations
             /// <summary>The 1.6 asset name.</summary>
             private const string NewAssetName = "Data/Characters";
 
+            /// <summary>The vanilla data without mod edits applied, used as the base when a pre-1.6 content pack loads the asset.</summary>
+            private readonly VanillaAssetFactory<Dictionary<string, CharacterData>> OriginalData = new(DataLoader.Characters);
+
 
             /*********
             ** Public methods
@@ -46,10 +49,13 @@ namespace ContentPatcher.Framework.Migrations
             /// <inheritdoc />
             public bool TryApplyLoadPatch<T>(LoadPatch patch, IAssetName assetName, [NotNullWhen(true)] ref T? asset, out string? error)
             {
-                Dictionary<string, string> tempData = patch.Load<Dictionary<string, string>>(this.GetOldAssetName(assetName));
-                Dictionary<string, CharacterData> newData = new();
-                this.MergeIntoNewFormat(newData, tempData, null);
-                asset = (T)(object)newData;
+                var data = this.OriginalData.GetFreshCopy();
+                var dataBackup = this.GetOldFormat(data);
+
+                var legacyLoad = patch.Load<Dictionary<string, string>>(this.GetOldAssetName(assetName));
+                this.MergeIntoNewFormat(data, legacyLoad, dataBackup);
+
+                asset = (T)(object)data;
 
                 error = null;
                 return true;