+++ ContentPatcher/Framework/Migrations/Migration_2_0.ForObjectInformation.cs
@@ -28,6 +28,9 @@ namespace ContentPatcher.Framework.Migrations
             /// <summary>The 1.6 asset name.</summary>
             private const string NewAssetName = "Data/Objects";
 
+            /// <summary>The vanilla data without mod edits applied, used as the base when a pre-1.6 content pack loads the asset.</summary>
+            private readonly VanillaAssetFactory<Dictionary<string, ObjectData>> OriginalData = new(DataLoader.Objects);
+
 
             /*********
             ** Public methods
@@ -47,10 +50,13 @@ namespace ContentPatcher.Framework.Migrations
             /// <inheritdoc />
             public bool TryApplyLoadPatch<T>(LoadPatch patch, IAssetName assetName, [NotNullWhen(true)] ref T? asset, out string? error)
             {
-                Dictionary<string, string> tempData = patch.Load<Dictionary<string, string>>(this.GetOldAssetName(assetName));
-                Dictionary<string, ObjectData> newData = new();
-                this.MergeIntoNewFormat(newData, tempData, null, patch.ContentPack.Manifest.UniqueID);
-                asset = (T)(object)newData;
+                var data = this.OriginalData.GetFreshCopy();
+                var dataBackup = this.GetOldFormat(data);
+
+                var legacyLoad = patch.Load<Dictionary<string, string>>(this.GetOldAssetName(assetName));
+                this.MergeIntoNewFormat(data, legacyLoad, dataBackup, patch.ContentPack.Manifest.UniqueID);
+
+                asset = (T)(object)data;
 
                 error = null;
                 return true;