+++ ContentPatcher/Framework/Migrations/Migration_2_0.ForWeapons.cs
@@ -21,6 +21,9 @@ namespace ContentPatcher.Framework.Migrations
             /// <summary>The asset name.</summary>
             private const string AssetName = "Data/Weapons";
 
+            /// <summary>The vanilla data without mod edits applied, used as the base when a pre-1.6 content pack loads the asset.</summary>
+            private readonly VanillaAssetFactory<Dictionary<string, WeaponData>> OriginalData = new(DataLoader.Weapons);
+
 
             /*********
             ** Public methods
@@ -40,10 +43,13 @@ namespace ContentPatcher.Framework.Migrations
             /// <inheritdoc />
             public bool TryApplyLoadPatch<T>(LoadPatch patch, IAssetName assetName, [NotNullWhen(true)] ref T? asset, out string? error)
             {
-                Dictionary<string, string> tempData = patch.Load<Dictionary<string, string>>(assetName);
-                Dictionary<string, WeaponData> newData = new();
-                this.MergeIntoNewFormat(newData, tempData, null);
-                asset = (T)(object)newData;
+                var data = this.OriginalData.GetFreshCopy();
+                var dataBackup = this.GetOldFormat(data);
+
+                var legacyLoad = patch.Load<Dictionary<string, string>>(assetName);
+                this.MergeIntoNewFormat(data, legacyLoad, dataBackup);
+
+                asset = (T)(object)data;
 
                 error = null;
                 return true;